Fix build failure when path of Dockerfile contains spaces.
Since the substitutions into the format call were not escaped, the shell
was interpreting paths with spaces in them as two arguments. Surrounding
said paths in spaces forces the shell to treat the literal spaces in the
path as part of the path instaead of a separator. Escaping spaces (" "
-> "\ ") would probably be a nicer solution, but the code is less
simple.