fix: false positives caused by auto-mode-alist pattern
The old pattern would match the following filenames:

- *Dockerfile
- *Dockerfile.*

This is because the pattern does not start with a slash indicating the
beginning of the basename. Personally this led to a few false positives,
like for example "siren-dockerfile.el" that sets up and configures
dockerfile-mode in my Emacs configuration.

This change restricts the patterns to:

- Dockerfile
- Dockerfile.*
- *.dockerfile

I believe this is still wide enough to capture all common naming
conventions for Dockerfiles.